Fix addInOrder
addInOrder generated tokens for unique words correctly, but
it modified the size of the list improperly in every case.

++size would be evaluated whenever a token was added, and
tokens which were added to the head or tail of the list,
rather than some point in between, would count as two tokens
(but only in respect to the size of the list). The size of
the list would be inflated, but the number of tokens would
be correct and the count of each token was still correct.

This propagated downstream when the size of the list was
used to iterate through the tokens in the list, printing
them.

Because the printing of tokensAlphabetical is dependent on
that list's size field, the inflated size does *not* trigger
`SLL.get()` to throw an IndexOutOfBoundsException, because
the sizes used are legal.

The while loop within `SLL.get()` tests if the next node
pointed to by the current node is null, and also if `i` is
greater than zero. The *intent* was to ensure that the iterative process of getting the next token in the list, while approaching the targeted index, didn't fall off the end of the list. That doesn't make sense, in hindsight, because the exception that may be thrown already handle's out of bounds issues. The loop runs as many times as it can, it approaches the pseudo-valid index and sets the token that will be returned to the last token in the list (due to the inflated `SLL.size` field), even though it hasn't reached a token that actually exists at such an index. That token is returned.

The printing process thus always prints the last token in the list the number of times the list was improperly inflated (which happens to be equal to the number of times a token was inserted at the head or tail of the list).